pyproject.toml
[project]
name = "multi-model-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "faiss-cpu>=1.13.0",
    "gradio>=6.0.1",
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.0",
    "numpy>=2.3.5",
    "pillow>=12.0.0",
    "pymupdf>=1.26.6",
    "scikit-learn>=1.7.2",
    "torch>=2.9.1",
    "transformers>=4.57.3",
]
